requirements.txt
ipykernel
langchain
langchain-community
langchain-huggingface
langchain-openai
langchain-core
transformers
youtube-transcript-api
openrouter
python-dotenv
streamlit
reportlab
sentence-transformers
faiss-cpu
langchain-groq
python-docx
docx2pdf
pypandoc
sounddevice
scipy
openai-whisper
gtts
langchain-text-splitters
pyttsx3